--- src/components/common/Layout/Layout.style.ts
@@ -1,18 +1,0 @@
-import styled from 'styled-components'
-
-const LayoutWrapper = styled.div`
-  display: flex;
-  align-content: center;
-
-  &.auth {
-    justify-content: center;
-
-    min-height: 100vh;
-  }
-
-  &.common {
-    justify-content: space-between;
-  }
-`
-
-export default LayoutWrapper
