--- src/components/common/Layout/Layout.tsx
@@ -1,10 +1,0 @@
-import { FC } from 'react'
-import { LayoutPropTypes } from './LayoutPropTypes'
-
-import LayoutWrapper from './Layout.style'
-
-const Layout: FC<LayoutPropTypes> = ({ content = 'common', children }) => {
-  return <LayoutWrapper className={content}>{children}</LayoutWrapper>
-}
-
-export default Layout
