--- src/components/common/Layout/LayoutPropTypes.ts
@@ -1,8 +1,0 @@
-import { ReactNode } from 'react'
-
-type content = 'auth' | 'common'
-
-export interface LayoutPropTypes {
-  content?: content
-  children: ReactNode
-}
