+++ src/components/common/UI/Button/EllipsisBtn/EllipsisBtn.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react'
 
-import { EllipsisBtnPropTypes } from './EllipsisBtnPropTypes'
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEllipsis } from '@fortawesome/free-solid-svg-icons'
+
+import { EllipsisBtnPropTypes } from './EllipsisBtnPropTypes'
 
 import EllipsisBtnWrapper from './EllipsisBtn.style'
