+++ src/components/common/UI/Container/ContainerPropTypes.ts
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 
-type flow = 'row' | 'column'
+type flow = 'auth' | 'row' | 'column'
 
 export interface ContainerPropTypes {
   children: ReactNode
