+++ src/components/common/UI/Form/Form/Form.style.ts
@@ -4,7 +4,6 @@ const FormWrapper = styled.form`
   max-width: 500px;
   width: 100%;
 
-  margin: auto 10px;
   padding: 20px;
 
   border-radius: 10px;
