+++ src/components/common/UI/Form/Input/Input.tsx
@@ -1,4 +1,5 @@
 import { FC } from 'react'
+
 import { InputPropTypes } from './InputPropTypes'
 
 import InputWrapper from './Input.style'
