+++ src/components/pages/Auth/Password/PasswordRecovery/PasswordRecovery.tsx
@@ -1,12 +1,13 @@
 import { FC } from 'react'
-import { useTypedSelector } from '../../../../../hooks/useTypedSelector'
-import { useActions } from '../../../../../hooks/useActions'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
+
+import { useTypedSelector } from '../../../../../hooks/useTypedSelector'
+import { useActions } from '../../../../../hooks/useActions'
+
 import { Input } from '../../../../../utils/types/input'
 import { recoverySchema } from '../../../../../utils/helpers/validationSchemas'
 
-import Layout from '../../../../common/Layout/Layout'
 import UI from '../../../../common/UI'
 
 const PasswordRecovery: FC = () => {
@@ -39,7 +40,7 @@ const PasswordRecovery: FC = () => {
   ]
 
   return (
-    <Layout content='auth'>
+    <UI.Container flow='auth'>
       <UI.Form.Container
         title='Восстановление пароля'
         tip='Введите Email, на который должна прийти ссылка для восстановления пароля'
@@ -64,7 +65,7 @@ const PasswordRecovery: FC = () => {
           </>
         }
       />
-    </Layout>
+    </UI.Container>
   )
 }
