+++ src/components/pages/Auth/Registration/Registration.tsx
@@ -1,12 +1,13 @@
 import { FC } from 'react'
-import { useTypedSelector } from '../../../../hooks/useTypedSelector'
-import { useActions } from '../../../../hooks/useActions'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
+
+import { useTypedSelector } from '../../../../hooks/useTypedSelector'
+import { useActions } from '../../../../hooks/useActions'
+
 import { Input } from '../../../../utils/types/input'
 import { signUpSchema } from '../../../../utils/helpers/validationSchemas'
 
-import Layout from '../../../common/Layout/Layout'
 import UI from '../../../common/UI'
 
 const Registration: FC = () => {
@@ -51,7 +52,7 @@ const Registration: FC = () => {
   ]
 
   return (
-    <Layout content='auth'>
+    <UI.Container flow='auth'>
       <UI.Form.Container
         title='Регистрация'
         tip='Для регистрации заполните все поля'
@@ -75,7 +76,7 @@ const Registration: FC = () => {
           </>
         }
       />
-    </Layout>
+    </UI.Container>
   )
 }
