+++ src/hooks/useTypedSelector.ts
@@ -1,4 +1,5 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux'
+
 import { RootState } from '../redux/reducers'
 
 export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector
