+++ src/utils/helpers/breakpoints.ts
@@ -1,5 +1,5 @@
 const size = {
-  mobileS: '320px',
+  // mobileS: '320px',
   mobileM: '375px',
   mobileL: '425px',
   tablet: '768px',
@@ -8,7 +8,7 @@ const size = {
 }
 
 export const device = {
-  mobileS: `(min-width: ${size.mobileS})`,
+  // mobileS: `(min-width: ${size.mobileS})`,
   mobileM: `(min-width: ${size.mobileM})`,
   mobileL: `(min-width: ${size.mobileL})`,
   tablet: `(min-width: ${size.tablet})`,
